Move SectionHeaderViewHolder.kt to mobile Also deleted SectionedAdapter.kt
