+++ mobile/src/main/java/com/boswelja/devicemanager/appmanager/ui/adapter/HeaderAdapter.kt
@@ -2,7 +2,7 @@ package com.boswelja.devicemanager.appmanager.ui.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.boswelja.devicemanager.common.recyclerview.item.SectionHeaderViewHolder
+import com.boswelja.devicemanager.common.ui.SectionHeaderViewHolder
 
 class HeaderAdapter(private val title: String) : RecyclerView.Adapter<SectionHeaderViewHolder>() {
